requirements.txt
altair==5.5.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
apturl==0.5.2
asgiref==3.9.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
bcrypt==3.2.0
blinker==1.8.2
boto3==1.35.4
botocore==1.35.4
Brlapi==0.8.3
cachetools==5.5.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.3.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dataclass-wizard==0.22.3
dbus-python==1.2.18
debugpy==1.8.6
decorator==5.1.1
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dnspython==2.6.1
duplicity==0.8.21
email_validator==2.2.0
eventlet==0.36.1
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.112.1
fasteners==0.14.1
filelock==3.6.0
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.55.4
future==0.18.2
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
gyp==0.1
h11==0.14.0
h5py==3.11.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka==1.3.5
kafka-python==2.0.2
keyring==23.5.0
kiwisolver==1.4.8
kombu==5.3.7
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mock==5.1.0
monotonic==1.6
more-itertools==8.10.0
nameko==2.14.1
narwhals==1.15.1
nest-asyncio==1.6.0
netifaces==0.11.0
numpy==1.26.4
numpy-quaternion==2023.0.4
oauthlib==3.2.0
olefile==0.46
orjson==3.10.12
packaging==24.1
pandas==2.2.3
paramiko==2.9.3
parso==0.8.4
path==16.14.0
path.py==12.5.0
pexpect==4.8.0
pillow==10.4.0
platformdirs==2.5.1
prompt_toolkit==3.0.48
protobuf==5.29.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==18.1.0
pyasn1==0.6.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
pymongo==4.6.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyquaternion==0.9.9
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.2.0
redis==5.0.8
referencing==0.35.1
reportlab==3.6.8
requests==2.32.3
rich==13.9.4
rpds-py==0.20.0
s3transfer==0.10.2
scipy==1.14.1
seaborn==0.13.2
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
spotipy==2.25.1
sqlparse==0.5.3
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.38.2
streamlit==1.40.2
systemd-python==234
tabulate==0.9.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.30.6
vine==5.1.0
virtualenv==20.13.0+ds
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
zod==0.5.0
